refactor: refactor root directive parsing
