+++ .gitignore
@@ -44,3 +44,4 @@ ubuntu_cgi_tester
 ubuntu_tester
 rfc*
 conf/noperm.conf
+YoupiBanane/
